Extract patch engine into reusable package
